cambio de int a double

+++ app/src/main/java/com/example/exoticworld/data/model/Productos.kt
@@ -4,6 +4,6 @@ data class Productos(
     val id: Int,
     val nombre: String,
     val descripcion: String,
-    val precio: Int,
+    val precio: Double,
     val categoriaId: Int
 )
